requirements.txt
PySide6>=6.4.0
pyqtgraph>=0.13.0
pyserial>=3.5
numpy>=1.23.0
